♻️ refactor(profile): move profile content into entities
