+++ src/entities/footprint/index.ts
@@ -1,0 +1,2 @@
+export { footprints } from "./model/footprints";
+export type { Footprint } from "./model/footprints";
