+++ src/entities/footprint/model/footprints.ts
