+++ src/entities/profile/index.ts
@@ -1,0 +1,2 @@
+export { links, techStacks } from "./model/profile";
+export { SocialLinks } from "./ui/social-links/social-links";
